change path to envs/kustomization.yaml for test stack (#31)
a conditional check to define the "Update apps" block based on the stack name. Specifically, for the TEST stack, we now use a different path to kustomization.yaml. This enhancement made the process more flexible when inheriting from the TEST stack to other environments like TEST2 and TEST3 and flexible control of apps