Fetch full history for sonar
When sonar has no access to base branch of PR, it analysis all the target branch issues, which basically blocks the PR if there any previous issues in the old code.

Fetching all the history with `fetch-depth: 0` is not the perfect solution (a lot of excess information which slows the build), but decent one for the time being.